requirements.txt
appnope==0.1.2
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
click==8.0.3
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.2
django-datatable==0.3.1
django-simple-search==1.0.2
et-xmlfile==1.1.0
executing==0.8.2
gunicorn==20.1.0
ipdb==0.13.9
ipython==8.0.1
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numpy==1.22.2
openpyxl==3.0.9
pandas==1.4.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
tqdm==4.62.3
traitlets==5.1.1
Unipath==1.1
wcwidth==0.2.5
Werkzeug==2.0.3
whitenoise==6.0.0
xlrd==2.0.1
